FALCON-1931 : Adding multiCluster tag for Multiple Cluster scenarios
In the regression test suite, there are lot of places multiCluster annotation has been defined to denote it uses multiple cluster. It is very useful annotation since we require additional cluster and setup to test the test case . Some of the major test cases such as feed replication and Hive DR are missing this annotation.

Author: Murali Ramasami <mramasami@Murali-Ramasami-MBP.local>
Author: Murali Ramasami <mramasami@hortonworks.com>

Reviewers: "Raghav Gautam <raghav@apache.org>"

Closes #123 from muraliramasami/multiClusterTag